backend/requirements.txt
# Core dependencies
numpy>=1.22.0
pandas>=1.4.0

# Computer Vision
opencv-python>=4.5.0

# Machine Learning
torch>=2.0.0
torchvision>=0.15.0

# API & Web Framework
fastapi>=0.95.0
uvicorn>=0.20.0
python-multipart>=0.0.5
pydantic>=2.0.0

# OpenAI
openai>=1.0.0

# Text-to-Speech
gTTS>=2.2.0
pyttsx3>=2.90

# Utilities
python-dotenv>=0.20.0
Pillow>=9.0.0
requests>=2.28.0
